ai-seo/requirements.txt
# AI SEO Canavari - Python Bagimliliklari

# Web Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0

# HTTP Client
httpx==0.28.1

# HTML Parsing & Crawling
beautifulsoup4==4.12.3
lxml==5.3.0

# Google APIs
google-api-python-client==2.159.0
google-auth==2.37.0
google-auth-httplib2==0.2.0

# LLM & AI
langchain==0.3.14
langchain-anthropic==0.3.4
langchain-openai==0.3.0
langchain-google-genai==2.0.8
langchain-core==0.3.30

# Database
sqlalchemy==2.0.36
psycopg2-binary==2.9.10

# Dashboard
streamlit==1.41.1

# Environment
python-dotenv==1.0.1

# Task Queue (opsiyonel)
# celery==5.4.0
# redis==5.2.1

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0

# Utilities
pydantic==2.10.4
